Adding hierarchy of measurements to tasks.
