+++ tests/test_metrics.py
@@ -17,7 +17,8 @@ import pytest
 from itertools import chain
 
 from wca.mesos import create_metrics, sanitize_label
-from wca.metrics import Metric, merge_measurements, MetricName, DerivedMetricName, METRICS_METADATA
+from wca.metrics import Metric, merge_measurements, MetricName, DerivedMetricName, METRICS_METADATA, \
+    export_metrics_from_measurements, MetricMetadata, MetricType, METRICS_LEVELS
 
 
 @pytest.mark.parametrize('label_key,expected_label_key', (
@@ -62,3 +63,42 @@ def test_merge_measurements(measurements_list, expected_merge):
 def test_metric_meta_exists():
     for metric_name in chain(MetricName, DerivedMetricName):
         assert metric_name in METRICS_METADATA, 'missing metadata for metric %s' % metric_name
+
+
+@pytest.mark.parametrize(
+    'measurements, expected', [
+        ({MetricName.MEM_NUMA_STAT_PER_TASK: {'1': 10, '2': 20}, }, 2),
+        ({MetricName.CYCLES: 1123}, 1),
+        ({}, 0)
+    ])
+def test_export_metrics_from_measurements(measurements, expected):
+    result = export_metrics_from_measurements('PLATFORM__', measurements)
+    assert len(result) == expected
+
+
+class TestMetric(object):
+    """To create and delete a test metric from metrics module structures."""
+    def __enter__(self):
+        MetricName.TEST_METRIC = 'test_metric'
+        METRICS_METADATA['test_metric'] = MetricMetadata(
+            MetricType.COUNTER, 'Non existing metric for unit test.')
+        METRICS_LEVELS['test_metric'] = ['numa_node', 'container']  # two levels
+
+    def __exit__(self, type, value, traceback):
+        """if exception was raised the metrics structeres will be cleaned up."""
+        del METRICS_METADATA['test_metric']
+        del METRICS_LEVELS['test_metric']
+        del MetricName.TEST_METRIC
+
+
+def test_export_metrics_from_measurements_artifical_metric():
+    """We currently do not have a metric which len(METRICS_LEVELS[X]) > 1,
+        so the need to add such metric in metrics structures for the test."""
+    with TestMetric():
+        measurements = {'test_metric': {'id0': {'stress': 0, 'dbmango': 1},
+                                        'id1': {'stress': 10, 'dbmango': 20}}}
+        result = export_metrics_from_measurements('PLATFORM__', measurements)
+        assert len(result) == 4
+        assert result[0].name == 'PLATFORM__test_metric'
+        assert 'numa_node' in result[0].labels and 'container' in result[0].labels
+        assert sorted([item.value for item in result]) == [0, 1, 10, 20]  # values for metrics
