+++ wca/platforms.py
@@ -26,7 +26,7 @@ import os
 from dataclasses import dataclass
 from enum import Enum
 
-from wca.metrics import Metric, MetricName, Measurements, MetricType
+from wca.metrics import Metric, MetricName, Measurements, MetricType, export_metrics_from_measurements
 from wca.profiling import profiler
 from wca import metrics
 
@@ -226,44 +226,6 @@ def get_platform_static_information():
     _platform_static_information['initialized'] = True
 
     return _platform_static_information
-
-
-def export_metrics_from_measurements(platform_prefix: str,
-                                     measurements: Measurements) -> List[Metric]:
-    all_metrics = []
-    for metric_name, metric_node in measurements.items():
-        if metric_name in metrics.METRICS_LEVELS:
-            levels = metrics.METRICS_LEVELS[metric_name]
-            max_depth = len(levels)
-
-            def is_leaf(depth):
-                return depth == max_depth
-
-            def create_metric(node, labels):
-                return [Metric.create_metric_with_metadata(
-                    name=platform_prefix + metric_name,
-                    value=node,
-                    labels=labels
-                )]
-
-            def recursive_create_metric(node, parent_labels=None, depth=0):
-                if is_leaf(depth):
-                    return create_metric(node, parent_labels)
-                else:
-                    metrics = []
-                    for parent_label_value, child in node.items():
-                        new_parent_labels = {} if parent_labels is None else dict(parent_labels)
-                        new_parent_labels[levels[depth]] = str(parent_label_value)
-                        metrics.extend(recursive_create_metric(child, new_parent_labels, depth+1))
-                    return metrics
-            all_metrics.extend(recursive_create_metric(metric_node, {}, 0))
-        else:
-            metric_value = metric_node
-            all_metrics.append(Metric.create_metric_with_metadata(
-                name=platform_prefix + metric_name,
-                value=metric_value,
-            ))
-    return all_metrics
 
 
 def create_metrics(platform: Platform) -> List[Metric]:
